genomic_app.py
import streamlit as st
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from Bio import SeqIO, SeqUtils
from Bio.SeqUtils import GC, molecular_weight

st.set_page_config(
    page_title="Genomic Analysis Toolkit",
    page_icon="🧬",
    layout="wide"
)

def main():
    st.title("🧬 Genomic Analysis Toolkit")
    
    st.sidebar.title("Navigation")
    app_mode = st.sidebar.selectbox(
        "Choose Analysis Type",
        ["Home", "Sequence Analysis", "Variant Analysis", "File Converter"]
    )
    
    if app_mode == "Home":
        show_home()
    elif app_mode == "Sequence Analysis":
        sequence_analysis()
    elif app_mode == "Variant Analysis":
        variant_analysis()
    elif app_mode == "File Converter":
        file_converter()

def show_home():
    st.markdown("""
    ## Welcome to the Genomic Analysis Toolkit!
    
    **Features:**
    - Sequence Analysis: Analyze FASTA files
    - Variant Analysis: Process VCF files  
    - File Converter: Convert between formats
    """)

def sequence_analysis():
    st.header("Sequence Analysis")
    uploaded_file = st.file_uploader("Upload FASTA file", type=['fasta', 'fa', 'fna'])
    
    if uploaded_file is not None:
        sequences = list(SeqIO.parse(uploaded_file, "fasta"))
        st.write(f"Found {len(sequences)} sequences")

def variant_analysis():
    st.header("Variant Analysis")
    uploaded_file = st.file_uploader("Upload VCF file", type=['vcf'])
    
    if uploaded_file is not None:
        st.success("VCF file uploaded successfully!")

def file_converter():
    st.header("File Converter")
    st.write("Convert between file formats")

if __name__ == "__main__":
    main()