requirements.txt
streamlit>=1.28.0
biopython>=1.81
pandas>=2.0.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.12.0
